Create post-install script to move GovUK assets from node_modules
Script automates step 4 of the GovUK setup.

-r flag is used to recursively copy entire directory
-f flag is used to overwrite existing files if found

Files in `public/assets` that do not exist in `node_modules/govuk-fronte
nd/govuk/assets/*` are untouched.

I considered adding `public/assets` to .gitignore, but thought we would
want to include the assets folder in version control for if we need to
add custom assets.